move methods to endpoints folder
